requirements.txt
discord.py
python-dotenv
speedtest-cli
requests
